requirements.txt
chainlit
google-generativeai==0.5.4
pandasai==2.2.5
llama-index==0.10.44
llama-index-experimental==0.1.3
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-llms-gemini
llama-index-llms-groq==0.1.4
llama-index-llms-ollama==0.1.5
llama-index-llms-openai==0.1.22
llama-index-llms-openai-like==0.1.3
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-multi-modal-llms-gemini